requirements.txt
tensorflow-cpu
numpy
matplotlib
scikit-learn
pillow
streamlit
streamlit-drawable-canvas
seaborn